ClipAiTest1/requirements.txt
pytube
openai-whisper
ffmpeg-python
flask-cors
openai